+++ hata/discord/http/rate_limit_groups.py
@@ -1281,7 +1281,7 @@ Group Details
     - Limit : `5`
     - Resets after : `20.0`
 
-- scheduled_event_users_get_chunk
+- scheduled_event_user_get_chunk
     - Endpoint : `/guilds/{guild_id}/scheduled-events/{scheduled_event_id}/users`
     - Method : `PATCH`
     - Required auth : `bot`
@@ -2006,7 +2006,7 @@ scheduled_event_create = RateLimitGroup(LIMITER_GUILD)
 scheduled_event_delete = RateLimitGroup(LIMITER_GUILD)
 scheduled_event_get = RateLimitGroup(LIMITER_GUILD)
 scheduled_event_edit = RateLimitGroup(LIMITER_GUILD)
-scheduled_event_users_get_chunk = RateLimitGroup(LIMITER_GUILD)
+scheduled_event_user_get_chunk = RateLimitGroup(LIMITER_GUILD)
 sticker_guild_get_all = RateLimitGroup.unlimited()
 sticker_guild_create = RateLimitGroup()
 sticker_guild_edit = RateLimitGroup()